// src/interfaces/places.ts
// Generated by https://quicktype.io

export interface PlacesResponse {
	type: string;
	licence: string;
	features: Feature[];
}

export interface Feature {
	type: string;
	properties: Properties;
	bbox: number[];
	geometry: Geometry;
}

export interface Geometry {
	type: string;
	coordinates: number[];
}

export interface Properties {
	place_id: number;
	osm_type: string;
	osm_id: number;
	place_rank: number;
	category: string;
	type: string;
	importance: number;
	addresstype: string;
	name: string;
	display_name: string;
}